feat(frontend): add incomming friend requests to notification page
