+++ frontend/src/components/transactions/TransactionNotifications.tsx
@@ -1,0 +1,164 @@
+import {
+  Typography,
+  List,
+  Button,
+  Divider,
+  capitalize,
+  Box,
+} from "@mui/material";
+import { useEffect, useState, Fragment } from "react";
+
+import TransactionData from "../../../../types/entity/data/TransactionData";
+import { TransactionDeleteResponseErrors } from "../../../../types/responses/TransactionDeleteResponse";
+import { TransactionPatchResponseErrors } from "../../../../types/responses/TransactionPatchResponse";
+import { useAppBar } from "../../contexts/AppBarProvider";
+import { useTemporaryNotifications } from "../../contexts/TemporaryNotificationsProvider";
+import { useAppDispatch, useAppSelector } from "../../hooks/redux";
+import { TransactionsFilter } from "../../store/transactions/helpers";
+import { selectAllTransactions } from "../../store/transactions/selectors";
+import {
+  getTransactions,
+  deleteTransactionRequest,
+  payTransactionRequest,
+} from "../../store/transactions/thunks";
+import { formatCurrency } from "../../utils/currency";
+import { getUsersFullName } from "../../utils/string";
+import ConfirmationDialog from "../common/ConfirmationDialog";
+import TransactionListItem from "./TransactionListItem";
+
+type ActionType = "decline" | "pay";
+
+export default function TransactionNotifications(): JSX.Element {
+  const dispatch = useAppDispatch();
+  const temporaryNotifications = useTemporaryNotifications();
+
+  const { setTitle } = useAppBar();
+
+  useEffect(() => {
+    setTitle("Notifications");
+  }, [setTitle]);
+
+  const transactions = useAppSelector(selectAllTransactions());
+
+  useEffect(() => {
+    dispatch(getTransactions(TransactionsFilter.OwedPaymentRequests));
+  }, [dispatch]);
+
+  const [selectedTransaction, setSelectedTransaction] =
+    useState<TransactionData | null>(null);
+  const [actionType, setActionType] = useState<ActionType>("decline");
+
+  const showConfirmation = (
+    actionType: ActionType,
+    transaction: TransactionData
+  ) => {
+    setSelectedTransaction(transaction);
+    setActionType(actionType);
+  };
+
+  const handleConfirm = async (
+    transaction: TransactionData | null,
+    action: ActionType
+  ) => {
+    if (!transaction) {
+      return;
+    }
+
+    const thunkFunction =
+      action === "decline" ? deleteTransactionRequest : payTransactionRequest;
+    try {
+      await dispatch(thunkFunction(transaction.id)).unwrap();
+      temporaryNotifications.enqueueNotification({
+        message: `Payment request ${action === "decline" ? "denied" : "paid"}!`,
+        severity: "success",
+      });
+    } catch (error) {
+      let notificationMessage = "An error has occured, please try again.";
+      if (error instanceof Error) {
+        temporaryNotifications.enqueueNotification({
+          message: notificationMessage,
+          severity: "error",
+        });
+        throw error;
+      }
+      const errors =
+        error instanceof Object
+          ? (error as
+              | TransactionPatchResponseErrors
+              | TransactionDeleteResponseErrors)
+          : undefined;
+
+      if ((errors as TransactionPatchResponseErrors | undefined)?.amount) {
+        // Safely assert since the if statement ensures amount exists on object
+        notificationMessage = (errors as TransactionPatchResponseErrors)
+          .amount!;
+      }
+      temporaryNotifications.enqueueNotification({
+        message: notificationMessage,
+        severity: "error",
+      });
+    }
+    setSelectedTransaction(null);
+  };
+
+  return (
+    <>
+      {transactions.length > 0 ? (
+        <>
+          <List sx={{ width: "100%" }}>
+            {transactions.map((transaction, i) => (
+              <Fragment key={transaction.id}>
+                <TransactionListItem
+                  transaction={transaction}
+                  actions={
+                    <>
+                      <Button
+                        onClick={() => showConfirmation("decline", transaction)}
+                      >
+                        Decline
+                      </Button>
+                      <Button
+                        onClick={() => showConfirmation("pay", transaction)}
+                      >
+                        Pay
+                      </Button>
+                    </>
+                  }
+                />
+                {i !== transactions.length - 1 && (
+                  <Divider variant="inset" component="li" />
+                )}
+              </Fragment>
+            ))}
+          </List>
+          <ConfirmationDialog
+            title={`${capitalize(actionType)} Transaction Request`}
+            onCancel={() => setSelectedTransaction(null)}
+            onConfirm={() => handleConfirm(selectedTransaction, actionType)}
+            open={selectedTransaction !== null}
+          >
+            {selectedTransaction && (
+              <Typography>
+                Are you sure you'd like to {actionType} the payment request from{" "}
+                {getUsersFullName(selectedTransaction.creator)} for{" "}
+                {formatCurrency(selectedTransaction.amount)}?
+              </Typography>
+            )}
+          </ConfirmationDialog>
+        </>
+      ) : (
+        <>
+          <Box pt={2}>
+            <Typography
+              variant="body2"
+              textAlign="center"
+              sx={{ color: "text.secondary" }}
+            >
+              You're all caught up!
+            </Typography>
+          </Box>
+        </>
+      )}
+    </>
+  );
+}
