+++ frontend/src/pages/account/FriendsPage.tsx
@@ -1,21 +1,10 @@
 import { Box, Tab, Tabs, Typography } from "@mui/material";
 import { useEffect, useState } from "react";
+
 import FriendsPanel from "../../components/friends/FriendsPanel";
 import SentFriendRequestsPanel from "../../components/friends/SentFriendRequestsPanel";
-
 import { useAppBar } from "../../contexts/AppBarProvider";
-
-/**
- * A utlity function to generate accessibility properties for an element
- * @param name The index of the tab
- * @returns an object of properties
- */
-function tabA11yProps(index: number) {
-  return {
-    id: `friends-tab-${index}`,
-    "aria-controls": `friends-tabpanel-${index}`,
-  };
-}
+import { tabA11yProps } from "../../utils/accessibility";
 
 export default function FriendsPage(): JSX.Element {
   const { setTitle } = useAppBar();
@@ -41,13 +30,13 @@ export default function FriendsPage(): JSX.Element {
           <Tab
             label="Friends"
             value={0}
-            {...tabA11yProps(0)}
+            {...tabA11yProps("friends", 0)}
             sx={{ flexGrow: 1 }}
           />
           <Tab
             label="Sent Requests"
             value={1}
-            {...tabA11yProps(1)}
+            {...tabA11yProps("friends", 1)}
             sx={{ flexGrow: 1 }}
           />
         </Tabs>
