+++ frontend/src/store/friends/thunks.ts
@@ -4,8 +4,23 @@ import FriendData from "../../../../types/entity/data/FriendData";
 import UserId from "../../../../types/entity/ids/UserId";
 import FriendsQueryRequest from "../../../../types/requests/FriendsQueryRequest";
 import FriendDeleteResponse from "../../../../types/responses/FriendDeleteReponse";
+import FriendPatchResponse from "../../../../types/responses/FriendPatchResponse";
 import FriendsResponse from "../../../../types/responses/FriendsResponse";
 import LocalFriendsApi from "../../api/local/LocalFriendsApi";
+
+export const acceptFriendship = createAsyncThunk(
+  "friends/acceptFriendship",
+  async (friendId: UserId, thunkAPI): Promise<FriendData> => {
+    const res = await LocalFriendsApi.acceptFriendship(friendId);
+    const resData: FriendPatchResponse = await res.json();
+
+    if (resData.errors || !resData.data) {
+      throw thunkAPI.rejectWithValue(resData.errors);
+    }
+
+    return resData.data;
+  }
+);
 
 export const getFriends = createAsyncThunk(
   "friends/getFriends",
